feat: refactor database handling and API architecture configuration
